(feat) Implements device permissions requests in AppDelegate, deleted CameraPermissionsVC.
